Multiplexer module for ISO communication
This module starts and SDP server and waits
for incoming V2G connections (TCP, and optionally TLS).
Upon connection, it reads the SupportedAppProtocol.req message
to determine if the EV supports ISO-20 or not.

If ISO-20 is supported, it transparently bridges the connection
to the ISO-20 module (using TLS->TCP).

If the EV does not support ISO-20, it bridges to the ISO-2 module
instead.

Both the required iso-2 and iso-20 modules only need TCP enabled,
but they require to accept e.g. PnC on TCP sessions (as the actual
communication to the EV may be TLS).
Also the need to listen on localhost (::1) instead of the actual
PLC device.

Signed-off-by: Cornelius Claussen <cc@pionix.de>